Fidelity level 1: sequential pre-execution
