+++ ledger/participant-state/kvutils/src/main/scala/com/daml/ledger/validator/preexecution/LedgerReaderWithFingerprints.scala
@@ -3,23 +3,21 @@
 
 package com.daml.ledger.validator.preexecution
 import com.daml.ledger.participant.state.kvutils.Fingerprint
-import com.daml.ledger.validator.LedgerStateAccess
+import com.daml.ledger.validator.LedgerStateOperations
 import com.daml.ledger.validator.LedgerStateOperations.{Key, Value}
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class LedgerStateAccessReaderWithFingerprints[LogResult](
-    ledgerStateAccess: LedgerStateAccess[LogResult],
+class LedgerReaderWithFingerprints[LogResult](
+    ledgerStateOperations: LedgerStateOperations[LogResult],
     valueToFingerprint: Option[Value] => Fingerprint)(implicit executionContext: ExecutionContext)
     extends LedgerStateReaderWithFingerprints {
 
   override def read(keys: Seq[Key]): Future[Seq[(Option[Value], Fingerprint)]] =
-    ledgerStateAccess.inTransaction { state =>
+    for {
+      readResult <- ledgerStateOperations.readState(keys)
+    } yield
       for {
-        readResult <- state.readState(keys)
-      } yield
-        for {
-          maybeValue <- readResult
-        } yield maybeValue -> valueToFingerprint(maybeValue)
-    }
+        maybeValue <- readResult
+      } yield maybeValue -> valueToFingerprint(maybeValue)
 }
